feat: add support for code posts (#4)
This adds support for a second blog with syntax highlighting. There are some duplicated sections in here, especially in `gatsby-node.js`, that I think it would be great to address, but I don’t want to block this on refactoring.

This also deprecates the `Figure` component (which is a gross hack) for an `Image` component that takes advantage of Markdown’s ability to replace standard Markdown images with `gatsby-image` components.